Python Scripts/requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.9.3
Brotli==1.0.9
bs4==0.0.1
bson==0.5.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
click==8.0.0
colorama==0.4.4
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
docopt==0.6.2
fake-useragent==0.1.11
Flask==2.0.0
Flask-Compress==1.9.0
future==0.18.2
getmac==0.8.2
GPUtil==1.4.0
httpserver==1.1.0
idna==2.10
itsdangerous==2.0.0
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.0
Jinja2==3.0.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.0
more-itertools==8.7.0
mysql-connector-python==8.0.24
netifaces==0.10.9
numpy==1.20.3
pandas==1.2.4
plotly==4.14.3
portend==2.7.1
protobuf==3.15.8
psdash==0.6.2
psutil==5.8.0
py-cpuinfo==8.0.0
pycparser==2.20
pymongo==3.11.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
requests==2.25.1
retrying==1.3.3
six==1.15.0
soupsieve==2.2.1
tabulate==0.8.9
tempora==4.0.2
Unidecode==1.2.0
urllib3==1.26.4
Werkzeug==2.0.0
WMI==1.5.1
zc.lockfile==2.0
